+++ src/aosm/azext_aosm/tests/latest/integration_tests/integration_test_mocks/cnf_mocks/nginxdemo-0.1.0.tgz
