+++ src/aosm/azext_aosm/tests/latest/integration_tests/integration_test_mocks/mock_input_templates/cnf_nsd_input_template.jsonc
