+++ src/aosm/azext_aosm/tests/latest/integration_tests/integration_test_mocks/vnf_mocks/ubuntu.vhd
