+++ src/aosm/azext_aosm/tests/latest/integration_tests/recording_processors.py
