+++ src/aosm/azext_aosm/tests/latest/integration_tests/test_aosm_core_vnf_build.py
